refactor!: use a setting to decide if we should map fields to relay types or not
`strawberry-django-plus` used to do that, but when migrating this
I thought it would be enough to check if the type itself inherits from
`relay.Node`.

It turns out that it is not, specially for filters as those doesn't
inherit from `relay.Node`.

This adds an option to allow the user to use `ID` or `GlobalID`
for their auto fields, meaning that someone that only uses `Node`
types can set `MAP_AUTO_ID_AS_GLOBAL_ID=True` to make sure that
it will only use `GlobalID` to map those.