refactor: Remove guardian ObjectPermissionChecker monkey patch
This is an old monkey patch that was introduced back when this code
used to live in `strawberry-django-plus`, with the intention of avoiding
too many queries to the database when fetching permissions.

This ended up causing an issue, as checking for a permission would
cache the result, and later when trying to modify the permissions and
checking those again would return the cached result instead.

This seems like to not be required anymore as well, based on the fact
that we have a lot of unit tests that test the permissioning stuff
with guardian, while also ensuring that the number of db requests
are known, and none presented any issues.