Add support for nested creation/update in mutations. This also has the benefit of consistently calling `full_clean()` before creating related instances.
This does remove the `get_or_create()` calls and instead uses `create` only. The expectation here is that `key_attr` could and should be used to indicate what field should be used as the unique identifier, and not something hard coded that could have unintended side effects when creating related instances that don't have unique constraints and expect new instances to always be created.