fix(optimizer): Avoid merging prefetches when using aliases
Merging querysets is usually fine, but when using an alias it might be
that one queryset is filtering for something, and the other is filtering
for something else, for the same field. That can lead to wrong
results being returned.

From now on, if a field is specified more than once with an alias,
the optimizer will skip it, possibly causing n+1 issues, but avoiding
wrong results (which is worse).

Fix #695